Add Check Count

+++ apps/api/src/services/SubdepartmentService.js
@@ -1,5 +1,6 @@
 
 import * as subDepartmentRepo from '../dataAccess/subdepartmentRepo.js';
+import * as scheduleRepo from '../dataAccess/scheduleRepo.js';
 import AppError from '../utils/AppError.js';
 import { getAllDocuments } from './queryService.js';
 
@@ -26,6 +27,8 @@ export const updateSubDepartmentById = async (id, data) => {
 };
 
 export const deleteSubDepartmentById = async (id) => {
+    const scheduleCountCheck = await scheduleRepo.countFiltered({ subDepartmentId: id });
+    if (scheduleCountCheck > 0) throw new AppError('sub-department has assigned schedules, cannot delete', 400);
     const subDepartment = await subDepartmentRepo.Delete(id);
     if (!subDepartment) throw new AppError('SubDepartment not found', 404);
     return subDepartment;
